Return non-zero value when exception occurs
When an exception occurs after starting a multi-threaded strategy, the exception was reported but at the
end aoflagger would nevertheless return a zero value. This is unfortunate for pipelines, because they
don't notice the error in that case. This is fixed by returning non-zero in caes of any exception.
Reported by W. Williams. (https://github.com/lofar-astron/prefactor/issues/211)